tokens are added to update and delete user

+++ routes/user.js
@@ -2,6 +2,7 @@ const express = require("express");
 const User = require("../models/User");
 const bcrypt = require("bcrypt");
 const router = express.Router();
+const auth = require("../verifyToken");
 
 router.get("/", async (req, res) => {
   try {
@@ -15,6 +16,7 @@ router.get("/", async (req, res) => {
     res.status(500).json(error);
   }
 });
+
 router.get("/:id", async (req, res) => {
   try {
     const user = await User.findById(req.params.id);
@@ -26,44 +28,51 @@ router.get("/:id", async (req, res) => {
     res.json(500).json(error);
   }
 });
-router.delete("/:id", async (req, res) => {
-  try {
-    const deleted_user = await User.findByIdAndDelete(req.params.id);
-    if (!deleted_user) {
-      return res.status(400).json({ message: "user not found" });
+
+router.delete("/:id", auth, async (req, res) => {
+  if (req.user.id === req.params.id) {
+    try {
+      await User.findByIdAndDelete(req.params.id);
+
+      res.status(200).json({
+        message: "user deleted",
+      });
+    } catch (error) {
+      res.status(500).json(error);
     }
-    res.status(200).json({
-      message: "user deleted",
+  } else {
+    res.status(403).json({
+      message: "you can update only your account",
     });
-  } catch (error) {
-    res.status(500).json(error);
   }
 });
-router.put("/:id", async (req, res) => {
-  const hashedPassword = await bcrypt.hash(req.body.password, 12);
-  if (req.body.password) {
-    req.body.password = hashedPassword;
-  }
-  try {
-    const user = await User.findById(req.params.id);
-    !user &&
-      res.status(403).json({
-        message: "you can update only your account",
+
+router.put("/:id", auth, async (req, res) => {
+  if (req.user.id === req.params.id) {
+    const hashedPassword = await bcrypt.hash(req.body.password, 12);
+    if (req.body.password) {
+      req.body.password = hashedPassword;
+    }
+    try {
+      const updateUser = await User.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      const { password, ...updatedInfo } = updateUser._doc;
+      res.status(200).json({
+        message: "user updated",
+        data: updatedInfo,
       });
-    const updateUser = await User.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    const { password, ...updatedInfo } = updateUser._doc;
-    res.status(200).json({
-      message: "user updated",
-      data: updatedInfo,
+    } catch (error) {
+      res.status(500).json(error);
+    }
+  } else {
+    res.status(403).json({
+      message: "you can update only your account",
     });
-  } catch (error) {
-    res.status(500).json(error);
   }
 });
 module.exports = router;
